+++ lib/virtus/attribute/symbol.rb
@@ -1,0 +1,21 @@
+module Virtus
+  class Attribute
+
+    # Symbol
+    #
+    # @example
+    #   class Product
+    #     include Virtus
+    #
+    #     attribute :code, Symbol
+    #   end
+    #
+    #   product = Product.new(:code => :red)
+    #
+    class Symbol < Object
+      primitive       ::Symbol
+      coercion_method :to_symbol
+
+    end # class Symbol
+  end # class Attribute
+end # module Virtus