+++ spec/unit/virtus/attribute/set/coerce_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+describe Virtus::Attribute::Set, '#coerce' do
+  subject { attribute.coerce(value) }
+
+  let(:attribute) { described_class.new(:colors) }
+
+  context 'with an Array' do
+    let(:value) { [ :foo, 'bar', 'bar', :foo ] }
+
+    it { should eql(Set[:foo, 'bar']) }
+  end
+end